requirements.txt
asgiref==3.8.1
django==5.1.2
joblib==1.2.0
jpype1==1.5.0
konlpy==0.6.0
lxml==5.3.0
numpy==2.1.2
packaging==24.1
pandas==2.2.3
pip==24.3.1
python-dateutil==2.9.0.post0
pytz==2024.2
scikit-learn==1.5.2
scipy==1.14.1
setuptools==65.5.0
six==1.16.0
sklearn-pandas==2.2.0
sqlparse==0.5.1
threadpoolctl==3.5.0
tzdata==2024.2